requirements.txt
flask==3.0.3
flask-cors==4.0.0
SQLAlchemy==2.0.34
Flask-Migrate==4.0.7
python-dotenv==1.0.1
opencv-python==4.10.0.84
numpy==2.1.1
email-validator==2.2.0
werkzeug==3.0.3
psycopg2-binary==2.9.7
supabase
gunicorn==21.2.0

